lndservices: move router_client.go into the new package
